refactor(a11y): move accessibility utils and extract announcer provider
