+++ app/layout.tsx
@@ -3,8 +3,7 @@ import type { Metadata } from "next";
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
 import { Analytics } from "@vercel/analytics/next";
-import AccessibilityAnnouncerProvider from "@/app/providers/AccessibilityAnnouncerProvider";
-
+import AccessibilityAnnouncerProvider from "@/providers/AccessibilityAnnouncerProvider";
 
 export const metadata: Metadata = {
   metadataBase: new URL("https://haemologix.in"),
